make it possible to download models from gitlab
